+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 ### Modified
 
 * `symbolT` and `nameT` are replaced in favor of actual optics.
+* `pconstantClass`, `isAdaClass`, `psingleValue'`, `passetClassValueOf'` can now take both
+  `Tagged unit AssetClass` and `AssetClass`.
 
 ## 3.10.1 -- 2022-10-11